Fix bug #490: Add point and keep shape (near bone influence) : signal aborted
It prevents to reverse manipulate bone influenced blinepoints as a whole.
It needs code to allow to distribute the changes to the link when received a blinepoint.
For the moment it is not possible to transform a whole blinepoint so it looks like
we need to modify each component one by one.