Added target png_spritesheet
png_spritesheet is a target to put all frames in a single png file.
The frames are drawn vertically below each other.

I think it's a good start for this feature request:
http://www.synfig.org/issues/thebuggenie/synfig/issues/222
Although not everything wished there is implemented.

It's only a modification of the target png, there is no way to set the
space between two frames or to draw them horizontally.

Anyway it's useful for me and, I'm sure, for many others.