Keyframe synchro
- Keyframes are now synchronized and it's just User Interface, no new functionnality ;-)
When you select a keyframe from the time track, the corresponding keyframe in
the treeview panel will be selected and reverse.

- Developing, a deadloop issue occured sometime moving a keyframe from/to extreme position
and should be resolved. Please, play a little with this feature to verify it's robust enough.

- Sometime, when re-ordering (moving a kf in the kf widget on the time track)
the keyframe treeview get in wired sort mode and that put in trouble the
kf synchro. You need to re order the kf treeview by the time colomn to arrange this
problem.
related issue : http://www.synfig.org/issues/thebuggenie/synfig/issues/481

- Due to focus issues, the first mouse click is not catched.