Make png target default for png extension.
Explanation: ext_book is a std::map so keys occur only once.
The three TARGET macros for cairo_png_trgt, png_trgt and png_trgt_spritesheet,
overwrite the ext_book value for the key "png" so the last target defined
is the one selected by default when target is deduced from extension.

With the commit we assure that for extension "png" we will use png_trgt
when the target is deduced from the png.