Always use "By Layer Default" blend method on startup.
This allows to avoid the confusions when (nebbie) user accidentaly changes the default blend method to something like "Divide". In this case the user will be wondering why nothing appears on the scren when he creates something. With this change the blend method will be reset to default value when application is restarted.