IMAGE_DIR hardcoded to /usr/local/share/pixmaps
There's code to allow for a custom IMAGE_DIR, and use "/usr/local/share/pixmaps" if none is specified, but no code to specify that IMAGE_DIR to be ${CMAKE_INSTALL_PREFIX}/share/pixmaps. This makes synfigstudio look not very pretty when it cannot find any pixmaps after being installed to a custom prefix.