fix crash when undoing creation of a moved/changed vertex of spline
It happened because the history tree rebuilds on every
undo/redo. It doesn't do it anymore.

Long chain:
The history tree row description would call to
Action::ValueDescSet::get_local_name() and it would
make a 'long' stack until it crashes, as follow:

> Action::ValueDescSet::get_local_name()
 > ValueDesc::get_description()
  > ValueDesc::is_exported()
   > ValueDesc::get_value_node()
    > LinkableValueNode::get_link()
AND FINALLY it crashes due to the invalid link index!

Test case:
* Create a polygon or spline.
* Click on a vertex to open context menu
* Select "Create Item"
* Move the new vertex
* Ctrl+Z (ok, vertex 'unmoved')
* Ctrl+Z (crash)