Fix crash in Spline Width and return of improper type in Spline converters (PR #1890)
Fix #1731

The pull request contains two commits:

First fixes crash in 'Spline Width' converter when you try to wrap it in 'Switch'. Similar to previously addressed in 'Spline Vertex' and 'Spline Tangent' converters.
Second addresses the regression introduced in this commit: a90d014
It was no longer possible to convert real values to 'Spline Tangent/Width' and angle to 'Spline Tangent'. The reason is that upon creation of converter the size of spline would be zero and code will always return ValueBase of vector type so 'Bad Connection' error would be thrown (expected real, got vector). See linked issue for screenshot of the error.