[Synfig Studio] fix mainwindow not showing when is offscreen
Synfig Studio restores the last position and size of Main Window
on launching.

If user had multiple monitors when closing Synfig last time he used it,
Main Window could be moved to a no-more existent monitor/area
and not be visible.

Some Window Managers (like on MS Windows case) does not prevents
opening a window on a invisible area.

On MS Windows, closing Synfig Studio app in a minified state
moves the window to position (-32000,-32000).

fix #1871