[Synfig Studio] don't let inner layer be duplicated twice
If action LayerDuplicate has among the layers to be duplicated a
layer named "a" that is 'inside' of a group layer (pastecanvas)
named "g". In case of "g" is a layer included in the action layer
list too, don't duplicate the "a" twice, ie. the group and
its contents - including "a" - are duplicated, but "a" won't
be duplicated again.

It would clash the GUID of both "a" duplicates, and that leads
to crash.
A way to avoid it in a general way would be too troublesome :P
Many hierarchy levels and special GUIDs to handle.

Regular (and advanced) user normally won't need it, anyway.

Full description and minor discussion here:
https://forums.synfig.org/t/proposal-of-change-of-behavior-duplicating-both-group-and-an-inner-layer/11823