ValueDescLink: remove useless code
if(iter_scalar*current_scalar < 0) is always false because
get_scalar() used to be 0,0 or 1.0 only

So, the comparison is useless and so this whole section as
it would only change the found_inverse to true - but it would
never actually do it.

As this was the last code that would use ValueDesc::get_scalar(),
this method can be safely removed.