[Studio] Fix dialog responses for user pressing Esc key
When user presses Esc key or closes the dialog by clicking on
the X (close) button in window decoration, the dialog does not
return "RESPONSE_CANCEL", but "RESPONSE_DELETE_EVENT".

When programmatically closed, the dialog returns with
"RESPONSE_NONE".

In such cases, the dialog is supposed to be canceled, but
that was not correctly interpreted everywhere.