fix rendering of empty rows inside contour when Invert is on
Though not used by current GUI tools, a Contour may be composed
of more than one polygon or bline-shape.

Let's assume we are rendering a shape like the character 'i':
it is a dot above a straight line segment (or a bit more elaborated
shape), right?
There is at least a pixel row separating them.

If we set 'invert' parameter of the contour that contains both
polygons (dot + straight line segment),
the row(s) between those shapes are not filled, but kept as if
'invert' was not set.