Fix distance values on settings
They are saved with current preferred unit, but units were ignored when loading.

Many tools ignore the unit system a preference value was saved - that may be different
of current App preference/setting when loaded...
So, for example, 1 pixel can become the gigantic 1u (60px).

Test it:

    In Preferences, assure the Unit System is 'Pixels'
    Select Bline Tool
    In Tool Options Panel, set Brush Size to 1.0 px
    Draw any Bline in Work Area
    Change to Normal Tool
    In Preferences, assure the Unit System is 'Units'
    (optional: close Synfig and reopen it)
    Select Bline Tool again
    Draw another Bline
    It's way thicker than before. If you check Tool Options Panel, you'll see Brush Size as 1.0u

Mentioned first here:
 #2372 (review)