Quick fix for zstreambuf
The problem was in `inflateInit2` incorrect initialization.
In this case we have no file header, quote from documentation:

"windowBits can also be –8..–15 for raw inflate. In this case, -windowBits
determines the window size. inflate() will then process raw deflate data,
not looking for a zlib or gzip header, not generating a check value, and
not looking for any check values for comparison at the end of the
stream. This is for use with other formats that use the deflate
ompressed data format such as zip."

Anyway this change should be tested before the merge.