Adjust the path when building with msvc only
This is done since some of the modules of synfig have their own
dependencies. When loading any of these modules, their dependencies
have to be in the path to be successfully loaded.
VCPKG has the great feature of copying all the dependencies dlls
right beside the built target so that the path doesn't have to be changed.
But the modules reside in "modules" dir, while they are loaded from
"bin" dir, so the "modules" dir has to be in the path.