fix: synfig studio was not saving synfigapp settings on quit
Steps to reproduce the problem:

1. Select Bline Tool
2. Change a parameter in Tool Options Panel (e.g. Opacity or Line Width)
3. Restart Synfig Studio
4. Select Bline Tool and see the change was not saved.

Reason:
Since #2497, we finally use Gtk::Application as base class for `App`.
`App` is destructed before `MainWindow` is closed (basically at start).
So, stuff done on quitting before merging #2497 was not done on the right time now.

As `signal-shutdown` is only called in primary instance of Gtk::Application,
it's the proper moment to delete and finish stuff on application exit.