fix: fixed broken synfig.dll debugging in MSVC build (#2764)
While testing the Synfig MSVC build, I noticed that debugging of
functions located in the synfig.dll library does not work.

Visual Studio shows that this is happening because the pdb file
(with debug information) cannot be loaded. Attempting to load
this file manually gives a strange error message:
"No matching symbol file found in this folder."
However, if we rebuild the synfig target, then everything started
working.

This is happens because the synfig library (synfig.dll) and the
synfig executable (synfig.exe) have the same name and located
in the same directory (bin on Windows).

There are several solutions here:
1) Change the name of the `synfig` executable to something else,
like `synfig-cli`.

2) Set different output path for library pdb file:
```
set_target_properties(synfig PROPERTIES PDB_OUTPUT_DIRECTORY "lib")
```

3) Rename target in CMake from `synfig` to `libsynfig`

I prefer the last solution, but I'm not sure which one is better.