fix: Right-clicking a layer on a just opened file shows empty menu (#2781)
The issue was introduced by PR #2352

When the list of selected layers was empty, then
clicking on the layer opened an empty menu.

Now the logic has changed:
If the layer is in the list of selected layers, then
actions are applied to them, if not, then the selection
is removed from them and the layer that was clicked
becomes selected (and actions applied to it).