refactor: vocabulary initialization moved to `init_children_vocab` method
Just to make sure it's not being used somewhere.
I added a test flag (already removed here) to see if
`get_children_vocab` can be used to fill a dictionary
 after it has already been initialized. This test is never fired.

This is done because `get_children_vocab()` always copies data,
even when you just want to read it, so the splitting logic allows
us to reduce memory copying in the future.