ci: added conventional commits style check for PR titles
The main feature of these PR is the ability to automatically start
the check after editing the title of the PR, so no need to restart
this check manually.

This is made possible by the GitHub Actions event system.
In this case, we use the `PR edited` event.

```
  pull_request:
    types: [opened, edited]
```

P.S.
PR title is extracted from GitHub Actions JSON file located at ${GITHUB_EVENT_PATH}

https://docs.github.com/en/actions/learn-github-actions/environment-variables
> GITHUB_EVENT_PATH
> The path to the file on the runner that contains the full event webhook payload.
> For example, /github/workflow/event.json.