initializing radius of DuckDrag_SmoothMove with correct radius
the radius was being saved and loaded correctly in class  `StateSmoothMove_Context` . And it seems that this class is responsible for setting the radaius of the `DuckDrag_SmoothMove` class as when the spin buttons signal changed fires it updates its radius. However the problem was when the `DuckDrag_SmoothMove` class is constructed its
radius is initialized with 1 and not with the value saved in the settings.