feat: plugins can now have a (static) setup dialog before running
Info is passed as JSON (dict of strings) as the 3rd or 4th argument
to the script (depending if it needs info about current time and
layer selection).

The UI file must have the same name as the script file, except by the
file extension, that must be ".ui".
Synfig Studio will load the widget (container or not) identified as
"dialog_contents" (i.e. its ID must be "dialog_contents").
Fields must _named_ (its "name", not its "id") as this will be the
dictionary keys.