feat: added embedded Python
This is an example of how easy it is to embed Python.

But this adds a lot of questions:
1) Do we need plugin plugin support for console Synfig?
If so, this will require separating plugins into GUI and console ones.
2) Should plugin system should be implemented as a Synfig module?
At first glance, this will look strange - libsynfig loads a module that is linked to libsynfig.
3) Do we need an API architecture or just export whatever the plugin author needs?
4) Best way to iterate over layers.
Do we need to export the entire Layer class or just part of it?