refactor: use RadioToolButton instead of ToggleToolButton for tools dock (#3012)
Multiple calls (4+) to Dock_Toolbox::change_state_() occur because ToggleToolButton::set_active(), inside Dock_Toolbox::set_active_state(), would cause clicked signals.

This commit reduces the number of calls to just 2. Once for the previous tool button and a second for the new tool button. This is the normal behavior of RadioToolButtons.