fix: crash when clicking on rows of "Library Panel"  (#3046)
Brief bug description:

Clicking on the value column of any of the canvases or valuebase nodes rows causes a crash
this happens because when they are clicked a cellrenderer is in the process of creation 
but it fails in an assertion of the canvas. Now anyways I think it makes no sense to 
click on either of these and have a cellrenderer that could be used to edit anything.
As far as I understand both of these row are just for the purpose of organization (under canvases 
lies the canvases and under valuebase nodes lies all the exported valuebases etc..).

(Canvases and ValueBase Nodes rows -- old pic)
![image](https://user-images.githubusercontent.com/100296264/230830833-957731e5-2e12-4a27-bf77-fd5a7d442bb5.png)

What I did:

Made the cell renderers for these two rows  to be inactive as they are  not needed.