refactor: rename method to get_link_description to avoid ambiguity
There were two overloaded methods, and all of their arguments have
default values:

* LinkableValueNode::get_description(bool show_exported_name = true)
* LinkableValueNode::get_description(int index = -1, bool show_exported_name = true)

The first one is an override of the `ValueNode::get_description` method.

There was an unnecessary ambiguity here. So I renamed the second method
to `LinkableValueNode::get_link_description()` to make it explicit, and
`index` argument is not optional now.