remove dependency of select tool on handle visibility
now whenever the layer is clicked its ducks are all stored and then the needed ones are used to do the transformations needed by the select tool. This was an important step as now we can just disable handle visibility altogether when using this tool and only use it as an under the hood way of applying transformations without the gui part of it.