fix(MotionBlur): undefined behavior when subsample range is 0
For linear subsampling type, there is an undefined behavior
when both subsample (start and end) amounts are 0.

Synfig version 1.0.2: *none* is drawn!
Synfig version 1.2.1: a pale gray rectangle is drawn
Synfig version 1.4.3: a black gray rectangle is drawn

Anyway, this commit fixes the case when both are 0, by drawing
the context without any motion blur effect.

Note:
If they (the subsample amounts) are equal, but both nonzero, the
canvas is normally drawn.
Behavior is not the same between 1.0.2 and >=1.2.1 though.
The motion trail has a different number of 'echoes':
1.0.2 has less than the next releases.

fix #3345