refactor: delete Layer::get_sub_renddesc() with 2 parameters
Nobody fills that vector of RendDesc with more than one item.
The vector is always initialized empty, and the filled with a
single item - and not in any loop that could push back more items.

Even it would have more than one item, only the first one pushed
would be used anyway.