fix: inverted interpolation of gradient since 1.3.11
AFAIK it only affects Plant, Mandelbrot and Metaballs layers.
All the others use CompiledGradient, that doesn't suffer the issue.

Gradient::operator() was getting the values in wrong direction
(except for the gradient start and end positions)
@ 0.0 - first color (OK)
@ 1.0 - last color (OK)
@ 0.3 - closer to last color instead of first one
          (it looks like how it should be if position was 0.7)
@ 0.7 - closer to last color instead of first one
          (it looks like how it should be if position was 0.3)

For 3-or-more-color gradients, each interval is reversed,
but the start and end color are correct...

So the solution was to create an intermediate gradients that
keep the start and end color points, but create swapped color
points for each pair of them.

Consider a 3-color gradient like this:

Color A @0.0 -> Color B @0.3 -> Color C @1.0

, where @ means the color point position.

It is transformed to:

Color A @0.0 -> Color B @0.0 -> Color A @0.3 -> Color C @0.3 -> Color B @1.0 -> Color C @1.0

And it works fine.

Tests were created to check this inversion.

Layer versions were bumped, so it won't depend anymore about the Canvas version on load.