feat(action): "Reset Layer Parameter" set it to its default value
A new action was added to context menu of layer parameters.
It is accessible by right-clicking on Parameter panel or a layer
handle in the Canvas area.

This action allows user to change the current value of a layer
parameter to the default one, as it was from a brand new layer.

By design, if the parameter is animated, it clears the animation
first and then resets the value.

If the parameter is connected to an exported value or linked to
another value node, it disconnects it first.