fix: halftones masks obey the transformations
This PR creates a new interface for rendering Transformation
for TaskFilterPixelSW (actually TaskHalftone2 & TaskHalftone3).

For better layer rendering, it lets the transformations of
upper layers be propagated to lower layers (in other words, pass through
this task), as TaskInterfaceTransformationPass does.
Differently of the latter, though, this new interface,
called TaskInterfaceTransformationGetAndPass, pass the transformation
matrix not only to lower (i.e. context) layer, but to this task too.

HalfTone2 and HalfTone3 layers create masks with shapes that are
affected by the upper transformations, and that is the reason to create
this new interface.
So, the lower layers are better rendered due to the propagation of the
upper transformations (more precise rendering), as the same time as the
filter mask is correctly build with such transformation info.