fix: modifying waypoints for a group layer no longer modifies sibling layers (#3312)
When a group layer is selected, the time track will show waypoints for child layers in the group. Modifying one of these waypoints (remove, move, or duplicate) was also pulling in waypoints of sibling layers of the group, not just child layers, if the sibling also happened to have a waypoint at the same time.

Reproduction steps:

    Create 2 circle layers, and animate the location of both circles so that both have waypoints on the same frame
    Put one of the circles into a group layer
    Select the group layer, and right click a waypoint in the time track. Note that it says "Remove 2 Waypoints", even though there's only one layer in the group
    Select Remove. Both circle layers are modified, even though only one of them is in the selected group.

I'm still a bit fuzzy on the relationship between canvases and group layers. I tested that waypoints for imported files still works, but I don't know if there's another way to create a canvas object that would need to be tested as well.