fix(bline tool): ignores double-clicking if without any created point (#3594)
On a BLine Tool, double-clicking is used to realize the creation of a
new spline.

When no spline being drawn, double-clicking incorrectly makes impossible
to draw a new spline - user then would have to press Esc key to be able
to draw one more spline (Esc is a shortcut used to cancel an ongoing
spline).

Here we make the tool state ignore the double-click in such case,
preventing this awkward situation.

Fix #3388